+++ src/plugins/implementations/index.ts
@@ -4,3 +4,4 @@ export * from './DefaultPrefixManager';
 export * from './DefaultLanguager';
 export * from './Paginator';
 export * from './Main';
+export * from './Jishaku';